Moves TextController from core-android to core-common.
And make TextController.control(EditText) an extension function.
