+++ workflow-ui/core-android/api/core-android.api
@@ -70,14 +70,8 @@ public final class com/squareup/workflow1/ui/SnapshotParcelsKt {
 	public static final fun toSnapshot (Landroid/os/Parcelable;)Lcom/squareup/workflow1/Snapshot;
 }
 
-public final class com/squareup/workflow1/ui/TextController {
-	public fun <init> ()V
-	public fun <init> (Ljava/lang/String;)V
-	public synthetic fun <init> (Ljava/lang/String;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public final fun control (Landroid/widget/EditText;)V
-	public final fun getOnTextChanged ()Lkotlinx/coroutines/flow/Flow;
-	public final fun getTextValue ()Ljava/lang/String;
-	public final fun setTextValue (Ljava/lang/String;)V
+public final class com/squareup/workflow1/ui/TextControllerControlEditTextKt {
+	public static final fun control (Lcom/squareup/workflow1/ui/TextController;Landroid/widget/EditText;)V
 }
 
 public final class com/squareup/workflow1/ui/ViewBindingViewFactory : com/squareup/workflow1/ui/ViewFactory {
