+++ workflow-ui/core-common/api/core-common.api
@@ -26,6 +26,15 @@ public final class com/squareup/workflow1/ui/Named : com/squareup/workflow1/ui/C
 	public fun toString ()Ljava/lang/String;
 }
 
+public final class com/squareup/workflow1/ui/TextController {
+	public fun <init> ()V
+	public fun <init> (Ljava/lang/String;)V
+	public synthetic fun <init> (Ljava/lang/String;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public final fun getOnTextChanged ()Lkotlinx/coroutines/flow/Flow;
+	public final fun getTextValue ()Ljava/lang/String;
+	public final fun setTextValue (Ljava/lang/String;)V
+}
+
 public abstract interface annotation class com/squareup/workflow1/ui/WorkflowUiExperimentalApi : java/lang/annotation/Annotation {
 }
